fix distance calculation which affects compactness

+++ maskslic/processing.pyx
@@ -13,7 +13,7 @@ import numpy as np
 cimport numpy as cnp
 
 cdef extern from "processing.h":
-    vector[int] get_mean_point_distance(vector[int] &, vector[int] &, vector[int] &)
+    vector[float] get_mean_point_distance(vector[int] &, vector[int] &, vector[int] &)
 
 def get_mpd(cnp.ndarray[int, ndim=1] x, cnp.ndarray[int, ndim=1] y, cnp.ndarray[int, ndim=1] z):
     """
